apps/api-server/src/services/mrz-reader/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
numpy<2.0.0
passporteye==2.2.0
python-multipart==0.0.6
Pillow==10.2.0
pdf2image==1.16.3
